refactor: split protocol kernel helpers
